FFMpegModule.cpp: SetUpdateStatusCallback -> change param list type to int64_t, int, int (02/23/2022 - 11:34)

+++ FFMPegModule/FFMpegModule.cpp
@@ -35,7 +35,7 @@ DLLEXPORT void Release()
 	}
 }
 
-DLLEXPORT void SetUpdateStatusCallback(void (__stdcall* callback)(int64_t hour, int64_t minute, int64_t second))
+DLLEXPORT void SetUpdateStatusCallback(void (__stdcall* callback)(int64_t hour, int minute, int second))
 {
 	if (_avCapture)
 		_avCapture->_callback = callback;
